+++ src/lib_mockup/local_services.ml
@@ -562,7 +562,7 @@ module Make (E : MENV) = struct
          Directory.empty
          (* /chains/<chain_id>/blocks/<block_id>/helpers/preapply/operations *)
          E.Block_services.S.Helpers.Preapply.operations
-         (fun ((_, chain), _block) () op_list ->
+         (fun ((_, chain), _block) params op_list ->
            with_chain ~caller_name:"preapply operations" chain (fun () ->
                let*! outcome =
                  let* proto_state = partial_construction ~cache:`Lazy () in
@@ -586,7 +586,7 @@ module Make (E : MENV) = struct
                  return (List.rev acc)
                in
                match outcome with
-               | Ok result -> Tezos_rpc.Answer.return result
+               | Ok result -> Tezos_rpc.Answer.return (params#version, result)
                | Error errs -> Tezos_rpc.Answer.fail errs)))
 
   let hash_op (shell, proto) =