+++ tezt/lib_tezos/RPC.ml
@@ -513,6 +513,16 @@ let post_chain_block_helpers_preapply_block ?(chain = "main") ?(block = "head")
     ["chains"; chain; "blocks"; block; "helpers"; "preapply"; "block"]
     Fun.id
 
+let post_chain_block_helpers_preapply_operations ?(chain = "main")
+    ?(block = "head") ?version ~data () =
+  let query_string = Query_arg.opt "version" Fun.id version in
+  make
+    ~query_string
+    ~data
+    POST
+    ["chains"; chain; "blocks"; block; "helpers"; "preapply"; "operations"]
+    Fun.id
+
 let post_chain_block_helpers_forge_operations ?(chain = "main")
     ?(block = "head") ~data () =
   make