+++ CHANGES.rst
@@ -150,6 +150,12 @@ Node
   metric, and renamed the ``octez_mempool_pending_prechecked`` one to
   ``octez_mempool_pending_validated``. (MR :gl:`!9137`)
 
+- Added version ``1`` to RPC ``POST ../helpers/preapply/operations``. It can be
+  used by calling the RPC with the parameter ``?version=1`` (default version is
+  still ``0``). Version ``1`` allows the RPC to output ``attestation``,
+  ``preattestation``, ``double_attestation_evidence`` and
+  ``double_preattestation_evidence`` kinds in the JSON result. (MR :gl:`!8891`)
+
 Client
 ------
 - Adding client commands to generate, open and verify a time-lock.